Separate front_hash_test due to GCC compiler bugs
